pyproject.toml
[tool.poetry]
name = "brainrot-backend"
version = "0.1.0"
description = ""
authors = ["shiva kharbanda: <shiva.kharbanda@daaslabs.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.1"
python-decouple = "^3.8"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
psycopg2 = "^2.9.10"
huggingface-hub = "^0.31.1"
django-markdownify = "^0.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
